Workaroround for NEXUS-5799.
NEXUS-5799 (https://issues.sonatype.org/browse/NEXUS-5799) may cause the
apt plugin to be run before the min plugin, meaning that
updateArtifactInfo won't do anything, but updateDocument and
populateArtifactInfo will since the fextension property is set after
updateArtifactInfo is executed. By checking the filename in the Document
instead of the fextension property in the ArtifactInfo, this issue is
avoided and updateArtifactInfo will execute properly.

I also changed from using the fextension property to using packaging,
which is more appropriate.