Added support for signing the repository.
Restructured how files are generated and introduced the files:
 * Packages
 * Release
 * Release.gpg
 * apt-sign.key.gpg

The first three are standard APT repository files and the last one is
the public key used for signing the armored Release.gpg file. The last
two files are only available if the APT: Configuration capability has
been added and configured properly.

Files are now cached for 5 seconds after creation to speed up generation
of the Release file which would otherwise trigger the generation of
Packages two times, the same goes for Release.gpg. Without caching, the
Packages file would be generated 5 times during a single APT sync since
both Release and Release.gpg needs to be downloaded as well as either
Packages or Packages.gz. Files are now generated by FileGenerators and
hooked in to the DebianFileManager, this also makes it easy to add other
files in the future.