Updating modification time of all stored files so APT doesn't HIT
APT fetches files by issuing conditional requests with the last
modification time of the requested file, since this is set to when the
file was created and then never updated, APT will get a HIT on all
subsequent updates and only retrieve a new version if the Nexus server
is restarted.

The changes in this changeset modifies this behaviour by updating the
modification time of Packages(.gz) and Release(.gpg) whenever a file
ending in .deb is stored, updated or deleted.